Workload to address application resiliency
We are adding this workload so that we can explore how to write
applications that uses AMQP java client library so that they can
gracefully handle connection loss. Typically these connetions loss
happens during upgrades.

This workload pushes sample applications to Cloud Foundry so that we can
test failure scenarios