Use separate connections for consumption andn publishing
Having separate connections for consumption will make the consumer more
resilient because they will not be blocked when RabbitMQ server raises a
memory or disk alarm.
We need to inject 2 ConnectionFactories and we also need to be sure that
consumers get the consumer connection and producer components only the
producer connection.