Support AMQPs
before this change, we initialized the ConnectionFactory
with just username, password, virtualHost using the corresponding setter
methods. However, it is better to use setUri instead because that will
also take care of setting up the appropiate port when it is not
explicitly set in the address. In case of AMQPS, it will default to 5671
and for AMQP, it will default to 5672.