Configure Spring ConnectionFactory with SSL/TLS
The default Spring ConnectionFactory built by Spring Cloud Connectors
does not configure the underlying Java ConnectionFactory which by
default will never validate the server certificate.
With this change, we are building ourselves the Java ConnectionFactory
and the Spring ConnectionFactory which reads RabbitProperties from the
environment. This allows us to fully customize both ConnectionFactory(s)
and as a consecuence enable SSL with server certificate validation and
hostname verification.